requirements.txt
pandas==1.5.3
numpy==1.24.4
PyYAML==6.0.1